docker-compose/api/requirements.txt
fastapi
uvicorn
pydantic[email]
fastapi-cache2
redis-om
